% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{beta_warp_i}
\alias{beta_warp_i}
\title{Apply Beta CDF warp to functional data}
\usage{
beta_warp_i(
  shape1 = 1,
  shape2 = 1,
  trunc_beta = FALSE,
  dat_person,
  argval = "density_grid",
  fvars,
  return_df = FALSE,
  inverse,
  grid_interp
)
}
\arguments{
\item{shape1, shape2}{Numeric scalar: shape parameter for the beta distribution.}

\item{trunc_beta}{Logical: indicating whether to use truncated beta functions from the \code{cascsim} package.}

\item{dat_person}{data.frame containing the functional data}

\item{argval}{String: the name of the variable in \code{dat_person} representing the density grid.}

\item{fvars}{String: the name of the variable in \code{dat_person} representing the functional covariate to warp.}

\item{return_df}{Logical: if TRUE, return the warped data as data frames; if FALSE, return as lists.}

\item{inverse}{Logical: if FALSE, apply the forward transformation; if TRUE, apply the inverse transformation.}

\item{grid_interp}{Character string: the type of interpolation method for grid transformation. Options are "loess" or "bspline".}
}
\value{
If return_df is TRUE, a list containing two data frames: 'warped_newgrid' with the warped data on the new grid, and 'warped_origgrid' with the warped data on the original grid. If return_df is FALSE, a list containing four elements: 'warped_curves_newgrid' with the warped data on the new grid, 'mat_density_newgrid' representing the new grid, 'warped_curves_origgrid' with the warped data on the original grid, and 'mat_density_origgrid' representing the original grid.
}
\description{
This function applies a Beta CDF warp function to functional data.
}
