% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{get_label}
\alias{get_label}
\title{Get the variable labels for a vector of column names}
\usage{
get_label(vars_names, varnames)
}
\arguments{
\item{vars_names}{A data frame with columns "variable" (column name) and "label"}

\item{varnames}{A character vector containing column names.}
}
\value{
The corresponding labels of the variables in \code{varnames}.
}
\description{
Get the variable labels for a vector of column names
}
\keyword{Internal}
