% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raft.R
\name{loglik_pen_i}
\alias{loglik_pen_i}
\title{RAFT Penalized Loss Function for an Individual}
\usage{
loglik_pen_i(
  bblid_i,
  reg_type,
  warp_params_i,
  pfr_obj = NULL,
  idvar,
  outcome_var,
  dat_orig_obs,
  argvals_orig_obs,
  fvars,
  penalty = list(loglik = TRUE, dist_temp = TRUE, dist_orig = FALSE, dist_alphas =
    FALSE, dist_warp = TRUE),
  lambdas = list(loglik = 1, dist_temp = 1, dist_orig = 1, dist_alphas = 1, dist_warp =
    1),
  penalize_type = "dist",
  template_ind,
  scale_init = NULL
)
}
\arguments{
\item{bblid_i}{String: the identifier for the individual whose data is being transformed.}

\item{reg_type}{Character string: the type of warp transformation. Options are "beta_cdf" or "poly_warp".}

\item{warp_params_i}{Numeric vector: warp parameters for the specified individual.}

\item{pfr_obj}{Object: the fitted model object, if available. Default is NULL.}

\item{idvar}{String: the name of the variable in \code{dat_orig_obs} representing the IDs.}

\item{outcome_var}{String: the name of the variable in \code{dat_orig_obs} representing the outcome variable.}

\item{dat_orig_obs}{data.frame containing the data to be warped, with one observation per row.}

\item{argvals_orig_obs}{String: the name of the variable in \code{dat_orig_obs} representing the original time grid.}

\item{fvars}{Character vector: the names of the functional covariates.}

\item{penalty}{List: a list specifying which penalties to apply. Default is list(loglik = TRUE, dist_temp = TRUE, dist_orig = FALSE, dist_alphas = FALSE, dist_warp = TRUE).}

\item{lambdas}{List: a list specifying the penalty weights. Default is list(loglik = 1, dist_temp = 1, dist_orig = 1, dist_alphas = 1, dist_warp = 1).}

\item{penalize_type}{Character string: the type of penalization. Options are "dist" for distance-based penalization or "corr" for correlation-based penalization. Default is "dist".}

\item{template_ind}{Numeric scalar: the index of the template curve. Default is 5.}

\item{scale_init}{Numeric scalar: the initial scale parameter for the likelihood model. Default is NULL.}
}
\value{
A numeric scalar representing the penalized loss function for the individual.
}
\description{
This function calculates the penalized loss function for an individual based on the specified warp parameters, penalty settings, and likelihood model.
}
