% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raft.R
\name{max_loglik_all}
\alias{max_loglik_all}
\title{RAFT Algorithm}
\usage{
max_loglik_all(
  warp_params,
  reg_type = "beta_cdf",
  idvar = "bblid",
  outcome_var = "y1",
  pfr_obj = NULL,
  fvars = c("mod_fun_dmn7"),
  dat_orig_obs,
  argvals_orig_obs = "density_grid",
  penalty = list(loglik = TRUE, dist_temp = TRUE, dist_orig = FALSE, dist_alphas =
    FALSE, dist_warp = TRUE),
  lambdas = list(loglik = 1, dist_temp = 1, dist_orig = 1, dist_alphas = 1, dist_warp =
    1),
  penalize_type = "dist",
  template_ind = NULL,
  opt_it = 50,
  scale_init = NULL
)
}
\arguments{
\item{warp_params}{List: a list of warp parameters for all individuals.}

\item{reg_type}{Character string: the type of warp transformation. Options are "beta_cdf" or "poly_warp".}

\item{idvar}{String: the name of the variable in \code{dat_orig_obs} representing the IDs.}

\item{outcome_var}{String: the name of the variable in \code{dat_orig_obs} representing the outcome variable.}

\item{pfr_obj}{Object: the fitted \code{pfr} model object, if available. Default is NULL.}

\item{fvars}{Character vector: the names of the functional covariates.}

\item{dat_orig_obs}{data.frame containing the data to be warped, with one observation per row.}

\item{argvals_orig_obs}{String: the name of the variable in \code{dat_orig_obs} representing the original time grid.}

\item{penalty}{List: a list specifying which penalties to apply. Default is list(loglik = TRUE, dist_temp = TRUE, dist_orig = FALSE, dist_alphas = FALSE, dist_warp = TRUE).}

\item{lambdas}{List: a list specifying the penalty weights. Default is list(loglik = 1, dist_temp = 1, dist_orig = 1, dist_alphas = 1, dist_warp = 1).}

\item{penalize_type}{Character string: the type of penalization. Options are "dist" for distance-based penalization or "corr" for correlation-based penalization. Default is "dist".}

\item{template_ind}{Numeric scalar: the index of the template curve. Default is 5.}

\item{opt_it}{Numeric scalar: the number of iterations for optimization. Default is 50.}

\item{scale_init}{Numeric scalar: the initial scale parameter for the likelihood model. Default is NULL.}
}
\value{
A list containing the old and new warp parameters, their data frames, the distance between old and new warps, and the loss values.
}
\description{
The RAFT algorithm maximizes the penalized likelihood function for all individuals to estimate the optimal warp parameters.
}
