% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{warp_beta}
\alias{warp_beta}
\title{Warping Functions based on the Beta Cumulative Distribution Function (CDF)}
\usage{
warp_beta(x, shape1, shape2, trunc_beta, inverse)
}
\arguments{
\item{x}{Numeric vector: input values on which to evaluate the function.}

\item{shape1, shape2}{Numeric scalars: shape parameters of the beta distribution.}

\item{trunc_beta}{Logical: indicating whether to use truncated beta functions from the \code{cascsim} package.}

\item{inverse}{Logical: indicating whether to use the CDF function or the inverse CDF (quantile) function.}
}
\value{
Numeric vector: warped values of x based on the Beta CDF.
}
\description{
This function generates a warping function based on the Beta distribution using the specified
shape parameters, witht the option to use truncated beta functions and/or inverse functions.
The resulting function can be either a quantile function (inverse cumulative distribution function)
or a cumulative distribution function.
}
