% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{warp_data_one}
\alias{warp_data_one}
\title{Wrapper function to warp data given specified warping functions and warp parameters.}
\usage{
warp_data_one(
  dat_orig_obs,
  reg_type = "beta_cdf",
  bblid_i,
  idvar = "bblid",
  warp_params_i = c(1, 1),
  fvars = c("mod_fun_dmn7"),
  inverse = FALSE
)
}
\arguments{
\item{dat_orig_obs}{data.frame containing the data to be warped, with one observation per row.}

\item{reg_type}{String: the type of warp transformation. Options are "beta_cdf" or "poly_warp".}

\item{bblid_i}{String: the ID of the individual whose data is being transformed.}

\item{idvar}{String: the name of the variable in \code{dat_orig_obs} representing the IDs.}

\item{warp_params_i}{Numeric vector: warp parameters for the specified individual.}

\item{fvars}{Character vector: the names of the functional covariates.}

\item{inverse}{Logical: if FALSE, apply the forward transformation; if TRUE, apply the inverse transformation.}
}
\value{
The warped data for the specified individual.
}
\description{
This function warps individual observations in a dataset using specified warp parameters.
}
