% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{warp_poly}
\alias{warp_poly}
\title{Warping Functions based on simple polynomials}
\usage{
warp_poly(x, exp, inverse)
}
\arguments{
\item{x}{Numeric vector: input values on which to evaluate the function.}

\item{exp}{Numeric scalar: the exponent for the warp transformation.}

\item{inverse}{Logical: indicating whether to use the forward transformation or the inverse transformation.}
}
\value{
NNumeric vector: warped values of x based on the polynomial warping.
}
\description{
This function applies a polynomial warp transformation to the input data.
}
